refac: remove css and move input component to /ui/
